requirements.txt
praw~=7.7.1
numpy~=1.26.2
pandas~=2.1.3
nltk~=3.8.1
num2words~=0.5.13
python-dotenv~=1.0.0
flask~=3.0.0
torch~=2.1.1+cu121
transformers~=4.35.2
scikit-learn~=1.3.2
tqdm~=4.66.1